transport: route OnymSDK ephemeral signer through a provider injected from Identity
The Nostr transport was importing OnymSDK and calling Common.* directly,
which violated the rule that OnymSDK is reachable only from the Identity
layer. Strip OnymSDK out of Transport/Nostr/ entirely: the seam now
exposes only `NostrSigner` + `NostrEphemeralSignerProvider` protocols.
The OnymSDK-backed `OnymNostrSigner` impl + `OnymNostrSignerProvider`
factory move to Identity/ next to IdentityRepository (the only other
OnymSDK consumer). NostrMessageTransport / NostrInboxTransport take the
provider via constructor injection and ask it for a fresh signer per
outgoing event — no behaviour change, just no SDK leak.
